# frontend/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Telemedicina API",
    "description": ":V",
    "license": {
      "name": "Apache 2.0"
    },
    "version": "v0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/pacientes/{pacienteId}/historia-clinica/registros-medicos/{id}": {
      "get": {
        "tags": [
          "registro-medico-controller"
        ],
        "operationId": "obtenerRegistroMedico",
        "parameters": [
          {
            "name": "pacienteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistroMedicoRes"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "registro-medico-controller"
        ],
        "operationId": "modificarRegistroMedico",
        "parameters": [
          {
            "name": "pacienteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroMedicoReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistroMedicoRes"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "registro-medico-controller"
        ],
        "operationId": "eliminarRegistroMedico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
    "/pacientes/{id}": {
      "get": {
        "tags": [
          "paciente-controller"
        ],
        "operationId": "obtenerPaciente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PacienteRes"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "paciente-controller"
        ],
        "operationId": "editarPaciente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PacienteReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PacienteRes"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "paciente-controller"
        ],
        "operationId": "eliminarPaciente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/especialidades/{id}": {
      "get": {
        "tags": [
          "especialidad-controller"
        ],
        "operationId": "obtenerEspecialidad_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Especialidad"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "especialidad-controller"
        ],
        "operationId": "editarEspecialidad",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EspecialidadReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Especialidad"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "especialidad-controller"
        ],
        "operationId": "eliminarEspecialidad",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/doctores/{id}": {
      "get": {
        "tags": [
          "doctor-controller"
        ],
        "operationId": "obtenerDoctor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorRes"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "doctor-controller"
        ],
        "operationId": "editarDoctor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorRes"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "doctor-controller"
        ],
        "operationId": "eliminarDoctor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/usuarios": {
      "get": {
        "tags": [
          "usuario-controller"
        ],
        "operationId": "obtenerUsuario",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "usuario-controller"
        ],
        "operationId": "crearUsuario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          }
        }
      }
    },
    "/pacientes": {
      "get": {
        "tags": [
          "paciente-controller"
        ],
        "operationId": "obtenerPacientes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PacienteRes"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "paciente-controller"
        ],
        "operationId": "crearPaciente",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PacienteReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PacienteRes"
                }
              }
            }
          }
        }
      }
    },
    "/pacientes/{pacienteId}/historia-clinica/registros-medicos": {
      "get": {
        "tags": [
          "registro-medico-controller"
        ],
        "operationId": "obtenerRegistrosMedicos",
        "parameters": [
          {
            "name": "pacienteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistroMedico"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "registro-medico-controller"
        ],
        "operationId": "crearRegistroMedico",
        "parameters": [
          {
            "name": "pacienteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroMedicoReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistroMedicoRes"
                }
              }
            }
          }
        }
      }
    },
    "/especialidades": {
      "get": {
        "tags": [
          "especialidad-controller"
        ],
        "operationId": "obtenerEspecialidad",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Especialidad"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "especialidad-controller"
        ],
        "operationId": "crearEspecialidad",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EspecialidadReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Especialidad"
                }
              }
            }
          }
        }
      }
    },
    "/doctores": {
      "get": {
        "tags": [
          "doctor-controller"
        ],
        "operationId": "obtenerDoctores",
        "parameters": [
          {
            "name": "doctorSpec",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SpecificationDoctor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorRes"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "doctor-controller"
        ],
        "operationId": "crearDoctor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorRes"
                }
              }
            }
          }
        }
      }
    },
    "/pacientes/{paciendId}/historia-clinica": {
      "get": {
        "tags": [
          "historia-clinica-controller"
        ],
        "operationId": "obtenerHistorial",
        "parameters": [
          {
            "name": "paciendId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoriaClinicaRes"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RegistroMedicoReq": {
        "required": [
          "diagnostico",
          "doctorId",
          "fecha",
          "motivoDeLaConsulta",
          "observaciones",
          "seguimiento",
          "sintomas",
          "tratamiento"
        ],
        "type": "object",
        "properties": {
          "fecha": {
            "type": "string",
            "format": "date"
          },
          "doctorId": {
            "type": "integer",
            "format": "int64"
          },
          "motivoDeLaConsulta": {
            "type": "string"
          },
          "sintomas": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "diagnostico": {
            "type": "string"
          },
          "tratamiento": {
            "type": "string"
          },
          "seguimiento": {
            "type": "string"
          },
          "observaciones": {
            "type": "string"
          }
        }
      },
      "RegistroMedicoRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fecha": {
            "type": "string",
            "format": "date"
          },
          "doctorId": {
            "type": "integer",
            "format": "int64"
          },
          "motivoDeLaConsulta": {
            "type": "string"
          },
          "sintomas": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "diagnostico": {
            "type": "string"
          },
          "tratamiento": {
            "type": "string"
          },
          "seguimiento": {
            "type": "string"
          },
          "observaciones": {
            "type": "string"
          }
        }
      },
      "PacienteReq": {
        "type": "object",
        "properties": {
          "persona": {
            "$ref": "#/components/schemas/PersonaReq"
          }
        }
      },
      "PersonaReq": {
        "required": [
          "dni",
          "edad",
          "email",
          "nombre",
          "telefono"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "pattern": "^[a-zA-Z]+$",
            "type": "string"
          },
          "edad": {
            "type": "integer",
            "format": "int32"
          },
          "dni": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "PacienteRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "persona": {
            "$ref": "#/components/schemas/Persona"
          }
        }
      },
      "Persona": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "edad": {
            "type": "integer",
            "format": "int32"
          },
          "dni": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "EspecialidadReq": {
        "required": [
          "nombre"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "pattern": "^[a-zA-Z]+$",
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          },
          "fotoUrl": {
            "type": "string"
          }
        }
      },
      "Especialidad": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          },
          "fotoUrl": {
            "type": "string"
          }
        }
      },
      "DoctorReq": {
        "required": [
          "especialidadId",
          "honorarios",
          "licencia"
        ],
        "type": "object",
        "properties": {
          "persona": {
            "$ref": "#/components/schemas/PersonaReq"
          },
          "licencia": {
            "type": "string"
          },
          "honorarios": {
            "type": "number",
            "format": "double"
          },
          "especialidadId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "DoctorRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "persona": {
            "$ref": "#/components/schemas/Persona"
          },
          "licencia": {
            "type": "string"
          },
          "honorarios": {
            "type": "number",
            "format": "double"
          },
          "especialidadId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "rol": {
            "type": "string",
            "enum": [
              "DOCTOR",
              "PACIENTE"
            ]
          },
          "email": {
            "type": "string"
          }
        }
      },
      "Doctor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "persona": {
            "$ref": "#/components/schemas/Persona"
          },
          "licencia": {
            "type": "string"
          },
          "honorarios": {
            "type": "number",
            "format": "double"
          },
          "especialidad": {
            "$ref": "#/components/schemas/Especialidad"
          },
          "pacientes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Paciente"
            }
          }
        }
      },
      "HistoriaClinica": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "registroMedicos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistroMedico"
            }
          }
        }
      },
      "Paciente": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "persona": {
            "$ref": "#/components/schemas/Persona"
          },
          "historiaClinica": {
            "$ref": "#/components/schemas/HistoriaClinica"
          }
        }
      },
      "RegistroMedico": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "expediente": {
            "type": "string"
          },
          "fecha": {
            "type": "string",
            "format": "date"
          },
          "doctor": {
            "$ref": "#/components/schemas/Doctor"
          },
          "motivoDeLaConsulta": {
            "type": "string"
          },
          "sintomas": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "diagnostico": {
            "type": "string"
          },
          "tratamiento": {
            "type": "string"
          },
          "seguimiento": {
            "type": "string"
          },
          "observaciones": {
            "type": "string"
          }
        }
      },
      "HistoriaClinicaRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "registroMedicos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistroMedico"
            }
          }
        }
      },
      "SpecificationDoctor": {
        "type": "object"
      }
    }
  }
}